+++ src/main/java/mil/nga/wkb/util/GeometryUtils.java
@@ -5,6 +5,8 @@ import java.util.List;
 
 import mil.nga.wkb.geom.CircularString;
 import mil.nga.wkb.geom.CompoundCurve;
+import mil.nga.wkb.geom.Curve;
+import mil.nga.wkb.geom.CurvePolygon;
 import mil.nga.wkb.geom.Geometry;
 import mil.nga.wkb.geom.GeometryCollection;
 import mil.nga.wkb.geom.GeometryType;
@@ -17,6 +19,9 @@ import mil.nga.wkb.geom.Polygon;
 import mil.nga.wkb.geom.PolyhedralSurface;
 import mil.nga.wkb.geom.TIN;
 import mil.nga.wkb.geom.Triangle;
+import mil.nga.wkb.util.centroid.CentroidCurve;
+import mil.nga.wkb.util.centroid.CentroidPoint;
+import mil.nga.wkb.util.centroid.CentroidSurface;
 
 /**
  * Utilities for Geometry objects
@@ -25,6 +30,13 @@ import mil.nga.wkb.geom.Triangle;
  * @since 1.0.3
  */
 public class GeometryUtils {
+
+	/**
+	 * Default epsilon for line tolerance
+	 * 
+	 * @since 1.0.5
+	 */
+	public static final double DEFAULT_EPSILON = 0.000000000000001;
 
 	/**
 	 * Get the dimension of the Geometry, 0 for points, 1 for curves, 2 for
@@ -157,6 +169,11 @@ public class GeometryUtils {
 		case COMPOUNDCURVE:
 			minimize((CompoundCurve) geometry, maxX);
 			break;
+		case CURVEPOLYGON:
+			@SuppressWarnings("unchecked")
+			CurvePolygon<Curve> curvePolygon = (CurvePolygon<Curve>) geometry;
+			minimize(curvePolygon, maxX);
+			break;
 		case POLYHEDRALSURFACE:
 			minimize((PolyhedralSurface) geometry, maxX);
 			break;
@@ -272,6 +289,21 @@ public class GeometryUtils {
 	}
 
 	/**
+	 * Minimize the curve polygon
+	 * 
+	 * @param curvePolygon
+	 *            curve polygon
+	 * @param maxX
+	 *            max positive x value in the geometry projection
+	 */
+	private static void minimize(CurvePolygon<Curve> curvePolygon, double maxX) {
+
+		for (Curve ring : curvePolygon.getRings()) {
+			minimizeGeometry(ring, maxX);
+		}
+	}
+
+	/**
 	 * Minimize the polyhedral surface
 	 * 
 	 * @param polyhedralSurface
@@ -330,6 +362,11 @@ public class GeometryUtils {
 			break;
 		case COMPOUNDCURVE:
 			normalize((CompoundCurve) geometry, maxX);
+			break;
+		case CURVEPOLYGON:
+			@SuppressWarnings("unchecked")
+			CurvePolygon<Curve> curvePolygon = (CurvePolygon<Curve>) geometry;
+			normalize(curvePolygon, maxX);
 			break;
 		case POLYHEDRALSURFACE:
 			normalize((PolyhedralSurface) geometry, maxX);
@@ -461,6 +498,21 @@ public class GeometryUtils {
 
 		for (LineString lineString : compoundCurve.getLineStrings()) {
 			normalize(lineString, maxX);
+		}
+	}
+
+	/**
+	 * Normalize the curve polygon
+	 * 
+	 * @param curvePolygon
+	 *            curve polygon
+	 * @param maxX
+	 *            max positive x value in the geometry projection
+	 */
+	private static void normalize(CurvePolygon<Curve> curvePolygon, double maxX) {
+
+		for (Curve ring : curvePolygon.getRings()) {
+			normalizeGeometry(ring, maxX);
 		}
 	}
 
@@ -603,6 +655,442 @@ public class GeometryUtils {
 		double distance = Math.sqrt(Math.pow(x2 - x, 2) + Math.pow(y2 - y, 2));
 
 		return distance;
+	}
+
+	/**
+	 * Check if the point is in the polygon
+	 * 
+	 * @param point
+	 *            point
+	 * @param polygon
+	 *            polygon
+	 * @return true if in the polygon
+	 * @since 1.0.5
+	 */
+	public static boolean pointInPolygon(Point point, Polygon polygon) {
+		return pointInPolygon(point, polygon, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is in the polygon
+	 * 
+	 * @param point
+	 *            point
+	 * @param polygon
+	 *            polygon
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if in the polygon
+	 * @since 1.0.5
+	 */
+	public static boolean pointInPolygon(Point point, Polygon polygon,
+			double epsilon) {
+
+		boolean contains = false;
+		List<LineString> rings = polygon.getRings();
+		if (!rings.isEmpty()) {
+			contains = pointInPolygon(point, rings.get(0), epsilon);
+			if (contains) {
+				// Check the holes
+				for (int i = 1; i < rings.size(); i++) {
+					if (pointInPolygon(point, rings.get(i), epsilon)) {
+						contains = false;
+						break;
+					}
+				}
+			}
+		}
+
+		return contains;
+	}
+
+	/**
+	 * Check if the point is in the polygon ring
+	 * 
+	 * @param point
+	 *            point
+	 * @param ring
+	 *            polygon ring
+	 * @return true if in the polygon
+	 * @since 1.0.5
+	 */
+	public static boolean pointInPolygon(Point point, LineString ring) {
+		return pointInPolygon(point, ring, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is in the polygon ring
+	 * 
+	 * @param point
+	 *            point
+	 * @param ring
+	 *            polygon ring
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if in the polygon
+	 * @since 1.0.5
+	 */
+	public static boolean pointInPolygon(Point point, LineString ring,
+			double epsilon) {
+		return pointInPolygon(point, ring.getPoints(), epsilon);
+	}
+
+	/**
+	 * Check if the point is in the polygon points
+	 * 
+	 * @param point
+	 *            point
+	 * @param points
+	 *            polygon points
+	 * @return true if in the polygon
+	 * @since 1.0.5
+	 */
+	public static boolean pointInPolygon(Point point, List<Point> points) {
+		return pointInPolygon(point, points, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is in the polygon points
+	 * 
+	 * @param point
+	 *            point
+	 * @param points
+	 *            polygon points
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if in the polygon
+	 * @since 1.0.5
+	 */
+	public static boolean pointInPolygon(Point point, List<Point> points,
+			double epsilon) {
+
+		boolean contains = false;
+
+		int i = 0;
+		int j = points.size() - 1;
+		if (closedPolygon(points)) {
+			j = i++;
+		}
+
+		for (; i < points.size(); j = i++) {
+			Point point1 = points.get(i);
+			Point point2 = points.get(j);
+
+			// Shortcut check if polygon contains the point within tolerance
+			if (Math.abs(point1.getX() - point.getX()) <= epsilon
+					&& Math.abs(point1.getY() - point.getY()) <= epsilon) {
+				contains = true;
+				break;
+			}
+
+			if (((point1.getY() > point.getY()) != (point2.getY() > point
+					.getY()))
+					&& (point.getX() < (point2.getX() - point1.getX())
+							* (point.getY() - point1.getY())
+							/ (point2.getY() - point1.getY()) + point1.getX())) {
+				contains = !contains;
+			}
+		}
+
+		if (!contains) {
+			// Check the polygon edges
+			contains = pointOnPolygonEdge(point, points);
+		}
+
+		return contains;
+	}
+
+	/**
+	 * Check if the point is on the polygon edge
+	 * 
+	 * @param point
+	 *            point
+	 * @param polygon
+	 *            polygon
+	 * @return true if on the polygon edge
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnPolygonEdge(Point point, Polygon polygon) {
+		return pointOnPolygonEdge(point, polygon, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is on the polygon edge
+	 * 
+	 * @param point
+	 *            point
+	 * @param polygon
+	 *            polygon
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if on the polygon edge
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnPolygonEdge(Point point, Polygon polygon,
+			double epsilon) {
+		return polygon.numRings() > 0
+				&& pointOnPolygonEdge(point, polygon.getRings().get(0), epsilon);
+	}
+
+	/**
+	 * Check if the point is on the polygon ring edge
+	 * 
+	 * @param point
+	 *            point
+	 * @param ring
+	 *            polygon ring
+	 * @return true if on the polygon edge
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnPolygonEdge(Point point, LineString ring) {
+		return pointOnPolygonEdge(point, ring, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is on the polygon ring edge
+	 * 
+	 * @param point
+	 *            point
+	 * @param ring
+	 *            polygon ring
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if on the polygon edge
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnPolygonEdge(Point point, LineString ring,
+			double epsilon) {
+		return pointOnPolygonEdge(point, ring.getPoints(), epsilon);
+	}
+
+	/**
+	 * Check if the point is on the polygon ring edge points
+	 * 
+	 * @param point
+	 *            point
+	 * @param points
+	 *            polygon points
+	 * @return true if on the polygon edge
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnPolygonEdge(Point point, List<Point> points) {
+		return pointOnPolygonEdge(point, points, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is on the polygon ring edge points
+	 * 
+	 * @param point
+	 *            point
+	 * @param points
+	 *            polygon points
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if on the polygon edge
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnPolygonEdge(Point point, List<Point> points,
+			double epsilon) {
+		return pointOnPath(point, points, epsilon, !closedPolygon(points));
+	}
+
+	/**
+	 * Check if the polygon outer ring is explicitly closed, where the first and
+	 * last point are the same
+	 * 
+	 * @param polygon
+	 *            polygon
+	 * @return true if the first and last points are the same
+	 * @since 1.0.5
+	 */
+	public static boolean closedPolygon(Polygon polygon) {
+		return polygon.numRings() > 0
+				&& closedPolygon(polygon.getRings().get(0));
+	}
+
+	/**
+	 * Check if the polygon ring is explicitly closed, where the first and last
+	 * point are the same
+	 * 
+	 * @param ring
+	 *            polygon ring
+	 * @return true if the first and last points are the same
+	 * @since 1.0.5
+	 */
+	public static boolean closedPolygon(LineString ring) {
+		return closedPolygon(ring.getPoints());
+	}
+
+	/**
+	 * Check if the polygon ring points are explicitly closed, where the first
+	 * and last point are the same
+	 * 
+	 * @param points
+	 *            polygon ring points
+	 * @return true if the first and last points are the same
+	 * @since 1.0.5
+	 */
+	public static boolean closedPolygon(List<Point> points) {
+		boolean closed = false;
+		if (!points.isEmpty()) {
+			Point first = points.get(0);
+			Point last = points.get(points.size() - 1);
+			closed = first.getX() == last.getX() && first.getY() == last.getY();
+		}
+		return closed;
+	}
+
+	/**
+	 * Check if the point is on the line
+	 * 
+	 * @param point
+	 *            point
+	 * @param line
+	 *            line
+	 * @return true if on the line
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnLine(Point point, LineString line) {
+		return pointOnLine(point, line, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is on the line
+	 * 
+	 * @param point
+	 *            point
+	 * @param line
+	 *            line
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if on the line
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnLine(Point point, LineString line,
+			double epsilon) {
+		return pointOnLine(point, line.getPoints(), epsilon);
+	}
+
+	/**
+	 * Check if the point is on the line represented by the points
+	 * 
+	 * @param point
+	 *            point
+	 * @param points
+	 *            line points
+	 * @return true if on the line
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnLine(Point point, List<Point> points) {
+		return pointOnLine(point, points, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is on the line represented by the points
+	 * 
+	 * @param point
+	 *            point
+	 * @param points
+	 *            line points
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if on the line
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnLine(Point point, List<Point> points,
+			double epsilon) {
+		return pointOnPath(point, points, epsilon, false);
+	}
+
+	/**
+	 * Check if the point is on the path between point 1 and point 2
+	 * 
+	 * @param point
+	 *            point
+	 * @param point1
+	 *            path point 1
+	 * @param point2
+	 *            path point 2
+	 * @return true if on the path
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnPath(Point point, Point point1, Point point2) {
+		return pointOnPath(point, point1, point2, DEFAULT_EPSILON);
+	}
+
+	/**
+	 * Check if the point is on the path between point 1 and point 2
+	 * 
+	 * @param point
+	 *            point
+	 * @param point1
+	 *            path point 1
+	 * @param point2
+	 *            path point 2
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @return true if on the path
+	 * @since 1.0.5
+	 */
+	public static boolean pointOnPath(Point point, Point point1, Point point2,
+			double epsilon) {
+
+		boolean contains = false;
+
+		double x21 = point2.getX() - point1.getX();
+		double y21 = point2.getY() - point1.getY();
+		double xP1 = point.getX() - point1.getX();
+		double yP1 = point.getY() - point1.getY();
+
+		double dp = xP1 * x21 + yP1 * y21;
+		if (dp >= 0.0) {
+
+			double lengthP1 = xP1 * xP1 + yP1 * yP1;
+			double length21 = x21 * x21 + y21 * y21;
+
+			if (lengthP1 <= length21) {
+				contains = Math.abs(dp * dp - lengthP1 * length21) <= epsilon;
+			}
+		}
+
+		return contains;
+	}
+
+	/**
+	 * Check if the point is on the path between the points
+	 * 
+	 * @param point
+	 *            point
+	 * @param points
+	 *            path points
+	 * @param epsilon
+	 *            epsilon line tolerance
+	 * @param circular
+	 *            true if a path exists between the first and last point (a non
+	 *            explicitly closed polygon)
+	 * @return true if on the path
+	 */
+	private static boolean pointOnPath(Point point, List<Point> points,
+			double epsilon, boolean circular) {
+
+		boolean onPath = false;
+
+		int i = 0;
+		int j = points.size() - 1;
+		if (!circular) {
+			j = i++;
+		}
+
+		for (; i < points.size(); j = i++) {
+			Point point1 = points.get(i);
+			Point point2 = points.get(j);
+			if (pointOnPath(point, point1, point2, epsilon)) {
+				onPath = true;
+				break;
+			}
+		}
+
+		return onPath;
 	}
 
 }